refactor: VideoInformation 도메인 수정
