+++ android/app/src/main/java/com/on/turip/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.on.turip.di
 
 import com.on.turip.data.content.repository.DefaultContentRepository
-import com.on.turip.domain.contents.repository.ContentRepository
+import com.on.turip.domain.videoinfo.contents.repository.ContentRepository
 
 object RepositoryModule {
     val contentRepository: ContentRepository by lazy {
