+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/VideoInformation.kt
@@ -1,10 +1,8 @@
 package com.on.turip.domain.videoinfo.contents
 
-import com.on.turip.domain.videoinfo.TripCourse
+import com.on.turip.domain.videoinfo.Trip
 
 data class VideoInformation(
     val content: Content,
-    val tripDuration: TripDuration,
-    val tripPlaceCount: Int,
-    val tripCourses: List<TripCourse> = emptyList<TripCourse>(),
+    val trip: Trip,
 )
